+++ backend/src/test/java/corea/BackendApplicationTests.java
@@ -1,4 +1,4 @@
-package corea.backend;
+package corea;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
